foodbit_spawn_time becomes a double as opposed to an int.

+++ Assets/Scripts/Environment/Ether.cs
@@ -24,7 +24,7 @@ public class Ether : MonoBehaviour {
 	public double total_energy;
 	double energy;
 	double foodbit_energy;
- 	float foodbit_spawn_time;
+ 	double foodbit_spawn_time;
 	public int foodbit_count;
 	int fb_spawn_range;
 	int start_number_foodbits;
@@ -38,7 +38,7 @@ public class Ether : MonoBehaviour {
 		
 		total_energy = 			(double) settings.contents[name]["total_energy"];
 		foodbit_energy = 		(double) settings.contents	[name]["foodbit_energy"];
-		foodbit_spawn_time = 	(int) settings.contents	[name]["foodbit_spawn_time"];
+		foodbit_spawn_time = 	(double) settings.contents	[name]["foodbit_spawn_time"];
 		fb_spawn_range = 		(int) settings.contents	[name]["fb_spawn_range"];
 		start_number_foodbits = (int) settings.contents [name]["start_number_foodbits"];
